[USER] OT247-17 - Model and migration for user, created
