+++ database/models/user.js
@@ -1,0 +1,48 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class User extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      User.belongsTo(models.Role, { foreignKey: 'roleId' })
+      User.hasMany(models.Comment, { foreignKey: 'user_id' })
+    }
+  };
+  User.init({
+    firstName: {
+      type: DataTypes.STRING,
+      allowNull: false
+    },
+    lastName: {
+      type: DataTypes.STRING,
+      allowNull: false
+    },
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique: true
+    },
+    password: {
+      type: DataTypes.STRING,
+      allowNull: false
+    },
+    photo: {
+      type: DataTypes.STRING,
+      allowNull: true
+    },
+    roleId: DataTypes.INTEGER,
+    deletedAt: DataTypes.DATE
+  }, {
+    sequelize,
+    modelName: 'User',
+    paranoid: true,
+    timestamps: true
+  });
+  return User;
+};
